refactor duplicate code into a common js file
